Core: Moved feature_metadata.py to core
