+++ core/src/autogluon/core/features/types.py
@@ -1,5 +1,4 @@
 import logging
-import re
 from collections import defaultdict
 
 import numpy as np
