+++ tabular/src/autogluon/tabular/features/generators/binned.py
@@ -4,9 +4,10 @@ import logging
 import pandas as pd
 from pandas import DataFrame
 
+from autogluon.core.features.feature_metadata import R_INT, R_FLOAT, S_BINNED
+
 from .abstract import AbstractFeatureGenerator
 from .. import binning
-from ..feature_metadata import R_INT, R_FLOAT, S_BINNED
 from ..utils import get_smallest_valid_dtype_int
 
 logger = logging.getLogger(__name__)
