+++ tabular/src/autogluon/tabular/features/generators/bulk.py
@@ -4,8 +4,9 @@ from typing import List
 import pandas as pd
 from pandas import DataFrame
 
+from autogluon.core.features.feature_metadata import FeatureMetadata
+
 from .abstract import AbstractFeatureGenerator
-from ..feature_metadata import FeatureMetadata
 
 logger = logging.getLogger(__name__)
