+++ tabular/src/autogluon/tabular/features/generators/datetime.py
@@ -3,8 +3,9 @@ import logging
 import pandas as pd
 from pandas import DataFrame
 
+from autogluon.core.features.feature_metadata import R_DATETIME, S_DATETIME_AS_OBJECT
+
 from .abstract import AbstractFeatureGenerator
-from ..feature_metadata import R_DATETIME, S_DATETIME_AS_OBJECT
 
 logger = logging.getLogger(__name__)
