+++ tabular/src/autogluon/tabular/features/generators/drop_unique.py
@@ -2,8 +2,9 @@ import logging
 
 from pandas import DataFrame
 
+from autogluon.core.features.feature_metadata import FeatureMetadata, R_CATEGORY, R_OBJECT, S_TEXT
+
 from .abstract import AbstractFeatureGenerator
-from ..feature_metadata import FeatureMetadata, R_CATEGORY, R_OBJECT, S_TEXT
 
 logger = logging.getLogger(__name__)
