+++ tabular/src/autogluon/tabular/features/generators/fillna.py
@@ -3,8 +3,9 @@ import logging
 import numpy as np
 from pandas import DataFrame
 
+from autogluon.core.features.feature_metadata import R_OBJECT
+
 from .abstract import AbstractFeatureGenerator
-from ..feature_metadata import R_OBJECT
 
 logger = logging.getLogger(__name__)
