+++ tabular/src/autogluon/tabular/models/fasttext/fasttext_model.py
@@ -11,8 +11,7 @@ import numpy as np
 import pandas as pd
 
 from autogluon.core.constants import BINARY, MULTICLASS
-
-from ...features.feature_metadata import S_TEXT
+from autogluon.core.features.feature_metadata import S_TEXT
 
 from ..abstract.abstract_model import AbstractModel
 from ..abstract.model_trial import skip_hpo
