+++ tabular/src/autogluon/tabular/models/lgb/lgb_model.py
@@ -14,6 +14,7 @@ from autogluon.core.utils.savers import save_pkl
 from autogluon.core.utils import try_import_lightgbm
 from autogluon.core import Int, Space
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS
+from autogluon.core.features.feature_metadata import R_OBJECT
 
 from . import lgb_utils
 from .callbacks import early_stopping_custom
@@ -23,7 +24,6 @@ from .hyperparameters.searchspaces import get_default_searchspace
 from .lgb_utils import construct_dataset
 from ..abstract.abstract_model import AbstractModel
 from ..utils import fixedvals_from_searchspaces
-from ...features.feature_metadata import R_OBJECT
 
 warnings.filterwarnings("ignore", category=UserWarning, message="Starting from version")  # lightGBM brew libomp warning
 logger = logging.getLogger(__name__)
