+++ tabular/src/autogluon/tabular/models/rf/rf_model.py
@@ -10,8 +10,8 @@ from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
 
 from autogluon.core.constants import MULTICLASS, REGRESSION
 from autogluon.core.utils.exceptions import NotEnoughMemoryError, TimeLimitExceeded
+from autogluon.core.features.feature_metadata import R_OBJECT
 
-from ...features.feature_metadata import R_OBJECT
 from ..abstract.model_trial import skip_hpo
 from ..abstract.abstract_model import AbstractModel
 from ...features.generators import LabelEncoderFeatureGenerator
