+++ tabular/src/autogluon/tabular/models/tab_transformer/tab_transformer_model.py
@@ -5,16 +5,16 @@ import time
 
 import numpy as np
 import pandas as pd
-from autogluon.core.utils.loaders import load_pkl
 from tqdm import tqdm
+
+from autogluon.core.constants import BINARY, REGRESSION, MULTICLASS
+from autogluon.core.features.feature_metadata import R_OBJECT, S_TEXT_NGRAM, S_TEXT_AS_CATEGORY
+from autogluon.core.utils import try_import_torch
+from autogluon.core.utils.loaders import load_pkl
 
 from .hyperparameters.parameters import get_default_param
 from .hyperparameters.searchspaces import get_default_searchspace
 from ..abstract.abstract_model import AbstractNeuralNetworkModel
-from autogluon.core.utils import try_import_torch
-from ...features.feature_metadata import R_OBJECT, S_TEXT_NGRAM, S_TEXT_AS_CATEGORY
-from autogluon.core.constants import BINARY, REGRESSION, MULTICLASS
-
 
 logger = logging.getLogger(__name__)
