+++ tabular/src/autogluon/tabular/models/xgboost/xgboost_model.py
@@ -1,14 +1,15 @@
 import os
 import time
 import logging
+
 from autogluon.core.constants import BINARY, MULTICLASS, REGRESSION, SOFTCLASS, PROBLEM_TYPES_CLASSIFICATION
+from autogluon.core.features.feature_metadata import R_OBJECT
 from autogluon.core.utils import try_import_xgboost
 
 from . import xgboost_utils
 from .hyperparameters.parameters import get_param_baseline
 from .hyperparameters.searchspaces import get_default_searchspace
 from ..abstract.abstract_model import AbstractModel
-from ...features.feature_metadata import R_OBJECT
 
 logger = logging.getLogger(__name__)
